Config après avoir redressé les roues codeuses !

+++ asserv/config/params.h
@@ -107,7 +107,7 @@ PARAM(double, frontParMetreCodeurD)
 
 // Distance entre les roues codeuses en mm
 // !!! ATTENTION : redémarrer l'asserv pour prendre en compte les changements !!
-PARAM(int64_t, distRoues)
+PARAM(double, distRoues)
 
 // Nombre d'UO pour un tic de codeur
 // !!! ATTENTION : redémarrer l'asserv pour prendre en compte les changements !!
